refactor extract tweakpane logic to 'panel' folder
