+++ modules/browser/src/panel/index.ts
@@ -1,0 +1,2 @@
+export * from './blades';
+export * from './property-panel';
